# specification/src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  title: Hero API
  description: Hero Description
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Generated server url
tags:
  - name: Heroes Controller
paths:
  /heroes:
    get:
      tags:
        - Heroes Controller
      summary: Get all heroes
      operationId: getHeroes
      parameters:
        - name: name
          in: query
          required: false
          example: Tony Stark
          schema:
            type: string
        - name: character
          in: query
          required: false
          example: IRON MAN
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/HeroResource'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
    post:
      tags:
        - Heroes Controller
      summary: Create a new hero
      operationId: postHero
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HeroResource'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HeroResource'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
  /heroes/{id}:
    get:
      tags:
        - Heroes Controller
      summary: Get a hero by id
      operationId: getHero
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HeroResource'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
    delete:
      tags:
        - Heroes Controller
      summary: Delete a hero by id
      operationId: deleteHero
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
    patch:
      tags:
        - Heroes Controller
      summary: Update a hero by id
      operationId: patchHero
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HeroResource'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HeroResource'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: string
components:
  schemas:
    HeroResource:
      required:
        - character
        - name
      type: object
      properties:
        id:
          type: integer
          format: int32
          example: 1
        name:
          type: string
          example: Tony Stark
        character:
          type: string
          example: IRON MAN